requirements.txt
pandas>=2.2.3
plotly>=6.1.1
requests>=2.32.3
scikit-learn>=1.6.1
scipy>=1.13.1
streamlit>=1.45.1
yfinance>=0.2.61
python-dotenv>=1.0.0
redis>=5.0.1
python-dateutil>=2.8.2
numpy>=1.24.0
